Make repos available in graphql resolvers
